requirements.txt
asgiref==3.2.3
beautifulsoup4==4.8.2
confusable-homoglyphs==3.2.0
dj-database-url==0.5.0
Django==1.11.28
django-appconf==1.0.3
django-bootstrap3==12.0.3
django-bootstrap4==1.1.1
django-decouple==2.1
django-imagekit==4.0.2
django-material==1.4.3
django-registration==2.4.1
django-tinymce==2.8.0
gevent==1.4.0
greenlet==0.4.15
gunicorn==20.0.4
pilkit==2.0
Pillow==5.0.0
pkg-resources==0.0.0
psycopg2==2.8.4
python-decouple==3.3
pytz==2019.3
six==1.14.0
soupsieve==2.0
SQLAlchemy==1.3.13
sqlparse==0.3.0
Werkzeug==1.0.0
whitenoise==5.0.1
